Add files via upload
This contains the files that were edited and coded in the Project 24 in WhitehatJR Pro Course. It has some physics based sprites with a sprite that moves with your mouse. You can do anything with those physics based sprite. It is coded in javascript using "matter.js". To change the logic, edit the code in the file 'sketch.js' file. It is the modified version of my "Project-24" repository.